Group method with other related methods
